feat: add deleteFormCandidates

+++ src/main/java/com/example/javacourse/requests/DeleteVacancyRequest.java
@@ -2,15 +2,17 @@ package com.example.javacourse.requests;
 
 import jakarta.persistence.criteria.CriteriaBuilder;
 
+import java.math.BigInteger;
+
 public class DeleteVacancyRequest {
-    private Long id;
+    private BigInteger id;
 
 
-    public Long getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
 }
